Cleanup: Move remaining functions into utilities
