+++ gizmos.py
@@ -1,20 +1,23 @@
+import math
+
 import bpy
 import bgl
 import gpu
 import blf
-
 from bpy.types import Gizmo, GizmoGroup
 from mathutils import Vector, Matrix
 from mathutils.geometry import intersect_point_line
 
-import math
-
-from . import functions, global_data, icon_manager
+from . import global_data, icon_manager
 from .model.types import SlvsDistance, SlvsAngle, SlvsDiameter
 from .declarations import GizmoGroups, Gizmos, Operators
 from .draw_handler import ensure_selection_texture
 from .utilities.constants import HALF_TURN, QUARTER_TURN
 from .utilities.math import pol2cart
+from .utilities.preferences import get_prefs
+from .utilities.draw import coords_arc_2d
+from .utilities.view import get_2d_coords, get_scale_from_pos
+from .utilities.index import rgb_to_index
 
 # NOTE: idealy gizmo would expose active element as a property and
 # operators would access hovered element from there
@@ -51,7 +54,7 @@ class VIEW3D_GT_slvs_preselection(Gizmo):
             bgl.glPixelStorei(bgl.GL_UNPACK_ALIGNMENT, 1)
             bgl.glReadPixels(mouse_x, mouse_y, 1, 1, bgl.GL_RGBA, bgl.GL_FLOAT, buffer)
         if buffer.to_list()[3] > 0:
-            index = functions.rgb_to_index(*buffer.to_list()[:-1])
+            index = rgb_to_index(*buffer.to_list()[:-1])
             if index != global_data.hover:
                 global_data.hover = index
                 context.area.tag_redraw()
@@ -86,7 +89,7 @@ class ConstraintGizmo:
     def _set_colors(self, context, constraint):
         """Overwrite default color when gizmo is highlighted"""
 
-        theme = functions.get_prefs().theme_settings
+        theme = get_prefs().theme_settings
         is_highlight = (
             constraint == global_data.highlight_constraint or self.is_highlight
         )
@@ -123,7 +126,7 @@ class VIEW3D_GT_slvs_constraint(ConstraintGizmo, Gizmo):
             if not entity or not hasattr(entity, "placement"):
                 return
 
-            pos = functions.get_2d_coords(context, entity.placement())
+            pos = get_2d_coords(context, entity.placement())
             if not pos:
                 return
 
@@ -205,7 +208,7 @@ class VIEW3D_GT_slvs_constraint_value(ConstraintGizmo, Gizmo):
         if not constr.visible or not hasattr(constr, "value_placement"):
             return
 
-        prefs = functions.get_prefs()
+        prefs = get_prefs()
         theme = prefs.theme_settings
         color = (
             theme.constraint.text_highlight
@@ -291,12 +294,12 @@ def get_overshoot(scale, dir):
     if dir == 0:
         return 0
     # use factor of 0.005 for one-half arrowhead
-    overshoot = scale * 0.005 * functions.get_prefs().arrow_scale
+    overshoot = scale * 0.005 * get_prefs().arrow_scale
     return -math.copysign(overshoot, dir)
 
 
 def get_arrow_size(dist, scale):
-    size = scale * 0.01 * functions.get_prefs().arrow_scale
+    size = scale * 0.01 * get_prefs().arrow_scale
     size = min(size, abs(dist * 0.67))
     size = math.copysign(size, dist)
     return size, size / 2
@@ -406,9 +409,7 @@ class VIEW3D_GT_slvs_distance(Gizmo, ConstraintGizmoGeneric):
             p1, p2 = p2, p1
         p1_global, p2_global = [self.matrix_world @ p for p in (p1, p2)]
 
-        scale_1, scale_2 = [
-            functions.get_scale_from_pos(p, rv3d) for p in (p1_global, p2_global)
-        ]
+        scale_1, scale_2 = [get_scale_from_pos(p, rv3d) for p in (p1_global, p2_global)]
 
         arrow_1 = get_arrow_size(half_dist, scale_1)
         arrow_2 = get_arrow_size(half_dist, scale_2)
@@ -504,7 +505,7 @@ class VIEW3D_GT_slvs_angle(Gizmo, ConstraintGizmoGeneric):
         scales = []
         lengths, widths = [], []  # Length is limited to no more than 1/3 the span
         for p in (p1, p2):
-            scale = functions.get_scale_from_pos(self.matrix_world @ p.to_3d(), rv3d)
+            scale = get_scale_from_pos(self.matrix_world @ p.to_3d(), rv3d)
             scales.append(scale)
 
             length = min(
@@ -527,7 +528,7 @@ class VIEW3D_GT_slvs_angle(Gizmo, ConstraintGizmoGeneric):
         if constr.text_inside():
             coords = (
                 *draw_arrow_shape(p1, p1 + p1_s, widths[0]),
-                *functions.coords_arc_2d(
+                *coords_arc_2d(
                     0, 0, radius, 32, angle=angle, offset=-half_angle, type="LINES"
                 ),
                 *draw_arrow_shape(p2, p2 + p2_s, widths[1]),
@@ -541,7 +542,7 @@ class VIEW3D_GT_slvs_angle(Gizmo, ConstraintGizmoGeneric):
             leader_length = leader_end - leader_start
             coords = (
                 *draw_arrow_shape(p1, p1 - p1_s, widths[0]),
-                *functions.coords_arc_2d(
+                *coords_arc_2d(
                     0,
                     0,
                     radius,
@@ -580,9 +581,7 @@ class VIEW3D_GT_slvs_diameter(Gizmo, ConstraintGizmoGeneric):
         p2 = pol2cart(dist, angle)
 
         p1_global, p2_global = [self.matrix_world @ p.to_3d() for p in (p1, p2)]
-        scale_1, scale_2 = [
-            functions.get_scale_from_pos(p, rv3d) for p in (p1_global, p2_global)
-        ]
+        scale_1, scale_2 = [get_scale_from_pos(p, rv3d) for p in (p1_global, p2_global)]
 
         arrow_1 = get_arrow_size(dist, scale_1)
         arrow_2 = get_arrow_size(dist, scale_2)
@@ -699,7 +698,7 @@ def constraints_mapping(context):
 
 
 def set_gizmo_colors(gz, failed):
-    theme = functions.get_prefs().theme_settings
+    theme = get_prefs().theme_settings
     color = theme.constraint.failed if failed else theme.constraint.default
     color_highlight = (
         theme.constraint.failed_highlight if failed else theme.constraint.highlight
@@ -791,7 +790,7 @@ class VIEW3D_GGT_slvs_constraint(GizmoGroup):
         return True
 
     def setup(self, context):
-        theme = functions.get_prefs().theme_settings
+        theme = get_prefs().theme_settings
 
         mapping = {}
         for c in context.scene.sketcher.constraints.all:
@@ -820,12 +819,12 @@ class VIEW3D_GGT_slvs_constraint(GizmoGroup):
                 gz.type = c.type
                 gz.index = context.scene.sketcher.constraints.get_index(c)
 
-                pos = functions.get_2d_coords(context, e.placement())
+                pos = get_2d_coords(context, e.placement())
 
                 gz.entity_index = e.slvs_index
 
                 ui_scale = context.preferences.system.ui_scale
-                scale = functions.get_prefs().gizmo_scale * ui_scale
+                scale = get_prefs().gizmo_scale * ui_scale
                 offset_base = Vector((scale * 1.0, 0.0))
                 offset = offset_base * i * ui_scale
