+++ model/base_entity.py
@@ -6,10 +6,14 @@ import bgl
 from bpy.props import IntProperty, StringProperty, BoolProperty
 from bpy.types import Context
 
-from .. import functions, global_data
+from .. import global_data
 from ..utilities import preferences
 from ..shaders import Shaders
 from ..declarations import Operators
+from ..utilities.preferences import get_prefs
+from ..utilities.index import index_to_rgb, breakdown_index
+from ..utilities.view import update_cb
+
 
 logger = logging.getLogger(__name__)
 
@@ -29,7 +33,7 @@ class SlvsGenericEntity:
         options={"SKIP_SAVE"},
     )
     fixed: BoolProperty(name="Fixed")
-    visible: BoolProperty(name="Visible", default=True, update=functions.update_cb)
+    visible: BoolProperty(name="Visible", default=True, update=update_cb)
     origin: BoolProperty(name="Origin")
     construction: BoolProperty(name="Construction")
     props = ()
@@ -90,7 +94,7 @@ class SlvsGenericEntity:
         return 20 * preferences.get_scale()
 
     def __str__(self):
-        _, local_index = functions.breakdown_index(self.slvs_index)
+        _, local_index = breakdown_index(self.slvs_index)
         return "{}({})".format(self.__class__.__name__, str(local_index))
 
     @property
@@ -166,7 +170,7 @@ class SlvsGenericEntity:
         return self.hover or self in global_data.highlight_entities
 
     def color(self, context: Context):
-        prefs = functions.get_prefs()
+        prefs = get_prefs()
         ts = prefs.theme_settings
         active = self.is_active(context.scene.sketcher.active_sketch)
         highlight = self.is_highlight()
@@ -247,7 +251,7 @@ class SlvsGenericEntity:
 
         bgl.glPointSize(self.point_size_select)
 
-        shader.uniform_float("color", (*functions.index_to_rgb(self.slvs_index), 1.0))
+        shader.uniform_float("color", (*index_to_rgb(self.slvs_index), 1.0))
         if not self.is_point():
             viewport = [context.area.width, context.area.height]
             shader.uniform_float("Viewport", viewport)
