+++ model/group_entities.py
@@ -8,8 +8,10 @@ from bpy.props import CollectionProperty
 from bpy.utils import register_classes_factory
 from mathutils import Vector, Euler
 
-from .. import functions, global_data
+from .. import global_data
 from ..utilities.constants import QUARTER_TURN
+from ..utilities.index import breakdown_index
+
 from .base_entity import SlvsGenericEntity
 from .utilities import slvs_entity_pointer, update_pointers
 from .point_3d import SlvsPoint3D
@@ -84,7 +86,7 @@ class SlvsEntities(PropertyGroup):
 
     @staticmethod
     def _breakdown_index(index: int):
-        return functions.breakdown_index(index)
+        return breakdown_index(index)
 
     @classmethod
     def recalc_type_index(cls, entity):
