+++ utilities/geometry.py
@@ -1,4 +1,6 @@
+import math
 from statistics import mean
+from typing import Tuple
 
 import mathutils
 from mathutils import Vector
@@ -23,3 +25,26 @@ def get_face_midpoint(quat, ob, face):
     pos = Vector((0, 0, dist))
     pos.rotate(quat)
     return ob.matrix_world @ pos
+
+
+def nearest_point_line_line(p1: Vector, d1: Vector, p2: Vector, d2: Vector) -> Vector:
+    n = d1.cross(d2)
+    n2 = d2.cross(n)
+    return p1 + ((p2 - p1).dot(n2) / d1.dot(n2)) * d1
+
+
+def line_abc_form(p1: Vector, p2: Vector) -> Tuple[float, float, float]:
+    a = p2.y - p1.y
+    b = p1.x - p2.x
+    return a, b, a * p1.x + b * p1.y
+
+
+def get_line_intersection(a1, b1, c1, a2, b2, c2) -> Vector:
+    det = a1 * b2 - a2 * b1
+    if det == 0:
+        # Parallel lines
+        return Vector((math.inf, math.inf))
+    else:
+        x = (b2 * c1 - b1 * c2) / det
+        y = (a1 * c2 - a2 * c1) / det
+        return Vector((x, y))
